+++ encoding/csv.ts
@@ -1,0 +1,2 @@
+/** @deprecated (will be removed after 0.170.0) use the extended API under std/encoding/csv/ instead */
+export * from "./csv/mod.ts"
